requirements.txt
requests
python-telegram-bot
beautifulsoup4
